requirements.txt
# Core dependencies for Adaptive Accessibility Tracker
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
deepface>=0.0.79
mediapipe>=0.10.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
tf-keras

# Optional dependencies for enhanced functionality
Pillow>=10.0.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0

# Development and testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Streamlit and Gemini
streamlit
python-dotenv
google-generativeai
beautifulsoup4
moviepy

# Wan 2.2
diffusers
transformers
accelerate
torch